chore: reorg into beets, read-write and struct
